整併前端可以依據類別取得產品以及烘焙方式

+++ noah-core/src/main/java/noah/core/dao/RoastTypeDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import noah.core.model.RoastType;
 
-public interface RoastTypeDao extends JpaRepository<RoastType, Long>{
+public interface RoastTypeDao extends JpaRepository<RoastType, Integer> {
 
 }
